# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Документация",
        "title": "ВИМБ API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/backup": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "- Создаем резервную копию MongoDВ. При пустых полях берет значения их config файла.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backup"
                ],
                "summary": "Создает backup MongoDB.",
                "operationId": "routes-post-backup",
                "parameters": [
                    {
                        "description": "Запрос",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mongodb_backup.SwaggerBackupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StreamResponse"
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                }
            }
        },
        "/api/v1/backup-list": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "- Возвращает список бэкапов, s3Key.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backup"
                ],
                "summary": "Список бэкапов MongoDB.",
                "operationId": "routes-post-backup-list",
                "parameters": [
                    {
                        "description": "Запрос",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mongodb_backup.SwaggerListBackupsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StreamResponse"
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                }
            }
        },
        "/api/v1/backup-restore": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "- Восстановление БД из резервной копии. Указать s3Key бэкапа.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backup"
                ],
                "summary": "Восстановление backup MongoDB.",
                "operationId": "routes-post-restore",
                "parameters": [
                    {
                        "description": "Запрос",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mongodb_backup.SwaggerRestoreRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StreamResponse"
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                }
            }
        },
        "/api/v1/budgets": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает данные по сделкам в разрезе канало-периодов.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Сделки"
                ],
                "summary": "Возвращает данные по сделкам в разрезе канало-периодов.",
                "operationId": "routes-get-budgets",
                "parameters": [
                    {
                        "description": "Запрос",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerGetBudgetsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StreamResponse"
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                }
            }
        },
        "/api/v1/budgets/load": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создание задач, на загрузку бюджетов, за выбранный период.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Сделки"
                ],
                "summary": "Создание задач, на загрузку бюджетов.",
                "operationId": "routes-load-budgets",
                "parameters": [
                    {
                        "description": "Запрос",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BudgetLoadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CommonResponse"
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                }
            }
        },
        "/api/v1/budgets/query": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Динамический запрос на загрузку сохраненных данных. Логические операторы: eq ==, ne !=, gt \u003e, lt \u003c, ge \u003e=, le \u003c=, in in, isnil is nil.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Сделки"
                ],
                "summary": "Загрузка сохраненных бюджетов.",
                "operationId": "routes-query-budgets",
                "parameters": [
                    {
                        "description": "Запрос",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BudgetQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CommonResponse"
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                }
            }
        },
...
        "/api/v1/spot": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Постановка спота в блок.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Споты"
                ],
                "summary": "Постановка спота в блок.",
                "operationId": "routes-add-spot",
                "parameters": [
                    {
                        "description": "Запрос",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerAddSpotRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StreamResponse"
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает список спотов.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Споты"
                ],
                "summary": "Возвращает список спотов.",
                "operationId": "routes-get-spots",
                "parameters": [
                    {
                        "description": "Запрос",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerGetSpotsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StreamResponse"
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаление спота в блоке.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Споты"
                ],
                "summary": "Удаление спота в блоке.",
                "operationId": "routes-delete-spot",
                "parameters": [
                    {
                        "description": "Запрос",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerDeleteSpotRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StreamResponse"
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.BudgetLoadRequest": {
            "type": "object",
            "properties": {
                "EndMonth": {
                    "type": "string",
                    "example": "201912"
                },
                "SellingDirectionID": {
                    "type": "string",
                    "example": "23"
                },
                "StartMonth": {
                    "type": "string",
                    "example": "201903"
                }
            }
        },
        "models.BudgetQuery": {
            "type": "object",
            "properties": {
                "AdtID": {
                    "type": "object",
                    "properties": {
                        "eq": {
                            "type": "integer",
                            "example": 700068653
                        }
                    }
                },
                "CnlID": {
                    "type": "object",
                    "properties": {
                        "eq": {
                            "type": "integer",
                            "example": 1020335
                        }
                    }
                },
                "Month": {
                    "type": "object",
                    "properties": {
                        "eq": {
                            "type": "integer",
                            "example": 201902
                        }
                    }
                }
            }
        },
...
        "models.CommonResponse": {
            "type": "object",
            "additionalProperties": true
        },
...
        "models.StreamResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "request": {
                    "type": "string"
                }
            }
        },
...
        "models.SwaggerAddSpotRequest": {
            "type": "object",
            "properties": {
                "AuctionBidValue": {
                    "type": "string"
                },
                "BlockID": {
                    "type": "string"
                },
                "FilmID": {
                    "type": "string"
                },
                "FixedPosition": {
                    "type": "string"
                },
                "Position": {
                    "type": "string"
                }
            }
        },
...
        "models.SwaggerDeleteSpotRequest": {
            "type": "object",
            "properties": {
                "SpotID": {
                    "type": "string"
                }
            }
        },
...
        "models.SwaggerGetBudgetsRequest": {
            "type": "object",
            "properties": {
                "AdvertiserList": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "ID": {
                                "type": "string"
                            }
                        }
                    }
                },
                "ChannelList": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "ID": {
                                "type": "string"
                            }
                        }
                    }
                },
                "EndMonth": {
                    "type": "string"
                },
                "SellingDirectionID": {
                    "type": "string"
                },
                "StartMonth": {
                    "type": "string"
                }
            }
        },
...
        "models.SwaggerGetSpotsRequest": {
            "type": "object",
            "properties": {
                "AdtList": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "AdtID": {
                                "type": "string"
                            }
                        }
                    }
                },
                "ChannelList": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "Cnl": {
                                "type": "string"
                            },
                            "Main": {
                                "type": "string"
                            }
                        }
                    }
                },
                "EndDate": {
                    "type": "string"
                },
                "InclOrdBlocks": {
                    "type": "string"
                },
                "SellingDirectionID": {
                    "type": "string"
                },
                "StartDate": {
                    "type": "string"
                }
            }
        },
...
        "mongodb_backup.SwaggerBackupRequest": {
            "type": "object",
            "properties": {
                "Db": {
                    "type": "string"
                },
                "Host": {
                    "type": "string"
                },
                "Password": {
                    "type": "string"
                },
                "Port": {
                    "type": "string"
                },
                "Username": {
                    "type": "string"
                }
            }
        },
        "mongodb_backup.SwaggerListBackupsRequest": {
            "type": "object"
        },
        "mongodb_backup.SwaggerRestoreRequest": {
            "type": "object",
            "properties": {
                "S3Key": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}